refactor: move session creation action
